Refactor so that StateUpdater doesn't need pages
